[NO-TASK] Fixed unit tests
